+++ packages/client-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/client-config
 
+## 1.5.3
+
+### Patch Changes
+
+- 72b2fe0: update aws-cdk lib to ^2.168.0
+- Updated dependencies [cfdc854]
+- Updated dependencies [72b2fe0]
+- Updated dependencies [65abf6a]
+- Updated dependencies [f6ba240]
+  - @aws-amplify/platform-core@1.3.0
+  - @aws-amplify/plugin-types@1.6.0
+
 ## 1.5.2
 
 ### Patch Changes